+++ vector/src/main/java/im/vector/riotx/features/roomprofile/settings/RoomSettingsFragment.kt
@@ -1,0 +1,84 @@
+/*
+ * Copyright 2020 New Vector Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package im.vector.riotx.features.roomprofile.settings
+
+import android.os.Bundle
+import android.view.View
+import androidx.core.view.isVisible
+import com.airbnb.mvrx.Loading
+import com.airbnb.mvrx.args
+import com.airbnb.mvrx.fragmentViewModel
+import com.airbnb.mvrx.withState
+import im.vector.matrix.android.api.util.toMatrixItem
+import im.vector.riotx.R
+import im.vector.riotx.core.extensions.cleanup
+import im.vector.riotx.core.extensions.configureWith
+import im.vector.riotx.core.extensions.observeEvent
+import im.vector.riotx.core.platform.VectorBaseFragment
+import im.vector.riotx.features.home.AvatarRenderer
+import im.vector.riotx.features.roomprofile.RoomProfileArgs
+import kotlinx.android.synthetic.main.fragment_room_member_list.*
+import kotlinx.android.synthetic.main.merge_overlay_waiting_view.*
+import javax.inject.Inject
+
+class RoomSettingsFragment @Inject constructor(
+        val viewModelFactory: RoomSettingsViewModel.Factory,
+        private val controller: RoomSettingsController,
+        private val avatarRenderer: AvatarRenderer
+) : VectorBaseFragment(), RoomSettingsController.Callback {
+
+    private val viewModel: RoomSettingsViewModel by fragmentViewModel()
+    private val roomProfileArgs: RoomProfileArgs by args()
+
+    override fun getLayoutResId() = R.layout.fragment_room_member_list
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        controller.callback = this
+        setupToolbar(roomMemberListToolbar)
+        recyclerView.configureWith(controller, hasFixedSize = true)
+        waiting_view_status_text.setText(R.string.please_wait)
+        waiting_view_status_text.isVisible = true
+
+        viewModel.requestErrorLiveData.observeEvent(this) {
+            displayErrorDialog(it)
+        }
+    }
+
+    override fun onDestroyView() {
+        recyclerView.cleanup()
+        super.onDestroyView()
+    }
+
+    override fun invalidate() = withState(viewModel) { viewState ->
+        controller.setData(viewState)
+        renderRoomSummary(viewState)
+    }
+
+    override fun onEnableEncryptionClicked() {
+        viewModel.handle(RoomSettingsAction.EnableEncryption)
+    }
+
+    private fun renderRoomSummary(state: RoomSettingsViewState) {
+        waiting_view.isVisible = state.currentRequest is Loading
+
+        state.roomSummary()?.let {
+            roomMemberListToolbarTitleView.text = it.displayName
+            avatarRenderer.render(it.toMatrixItem(), roomMemberListToolbarAvatarImageView)
+        }
+    }
+}