+++ vector/src/main/java/im/vector/riotx/features/settings/ignored/VectorSettingsIgnoredUsersFragment.kt
@@ -77,14 +77,6 @@ class VectorSettingsIgnoredUsersFragment @Inject constructor(
                 .show()
     }
 
-    private fun displayErrorDialog(throwable: Throwable) {
-        AlertDialog.Builder(requireActivity())
-                .setTitle(R.string.dialog_title_error)
-                .setMessage(errorFormatter.toHumanReadable(throwable))
-                .setPositiveButton(R.string.ok, null)
-                .show()
-    }
-
     // ==============================================================================================================
     // ignored users list management
     // ==============================================================================================================